test(ssg): move the pager flatten tests to their own file
`pagination/tests.rs` went past the 350-line limit for changed sources.
The flatten cases share one question — which entries the pager will step
onto — so they move together.
